feat: support backpressure
At present the FIFO queue that this module uses is unbounded which means
it can exhaust all available memory available to the process.

This PR adds `highWaterMark` and `objectMode` options similar to those
found in node streams to allow us to constrain memory usage.

When `objectMode` is `false` (the default), `highWaterMark` is
interpreted as the number of bytes that are allowed to be present in the
queue.  When `objectMode` is `true`, `highWaterMark` is the number of
objects that are allowed to be present in the queue.

If `objectMode` is `false` any attempt to `.push` a value that is not a
`Uint8Array` will throw an error.

Backpressure is introduced in that promises returned from calls to `.push`
will not return until there is space in the queue to accomodate them.

`.end` is also now async to ensure that the `{ done: true }` iterator
result is pushed into the queue after all `{ done: false, value: ... }`
results have been pushed.

It also swaps out the hand-rolled `Next` type for the built in
`IteratorResult` so `pushable`/`pushableV`s can extend `AsyncGenerator`.

BREAKING CHANGE: This module now expects `Uint8Array`s to be pushed by default, pass `objectMode: true` to push objects and if consuming from typescript, use generics to specify the type of object