feat: add readableLength property (#27)
Adds a `readableLength` property that tells you the current state of the queue.

Also adds an `objectMode` option similar to that found in node streams.

If `objectMode` is `true`, `readableLength` is the number of objects in the queue.

If `objectMode` is `false` (the default), only `Uint8Array`s can be passed to `.push` and `readableLength` is the total number of bytes in the queue.

BREAKING CHANGE: `Uint8Array`s are expected by default, pass `objectMode: true` to push any other data types. If using TypeScript, use generics to define the data type.