feat!: support backpressure
Incorporates the suggestion from #1 to support backpressure via
returning a promise from `.push` and `.end` methods.

Removes the `onEmpty` function as this was a quick hack to simulate
backpressure.

Adds a `.pushV` method to `pushableV` to guarantee all values make
it into the buffer in a single asynchronous operation.

Ports `p-fifo` to typescript and adds test suites from that and the
`fast-fifo` module.

BREAKING CHANGE: `.push` and `.end` now return promises